restructured acceptance tests
