+++ lib/src/main/java/truelayer/java/merchantaccounts/entities/transactions/MerchantAccountPayment.java
@@ -31,5 +31,7 @@ public class MerchantAccountPayment extends Transaction {
         AccountIdentifier accountIdentifier;
 
         String accountHolderName;
+
+        String userId;
     }
 }
