+++ lib/src/test/java/truelayer/java/acceptance/MerchantAccountsAcceptanceTests.java
@@ -1,0 +1,56 @@
+package truelayer.java.acceptance;
+
+import static truelayer.java.TestUtils.assertNotError;
+
+import java.util.List;
+import lombok.SneakyThrows;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import truelayer.java.http.entities.ApiResponse;
+import truelayer.java.merchantaccounts.entities.ListMerchantAccountsResponse;
+import truelayer.java.merchantaccounts.entities.transactions.Transaction;
+
+@DisplayName("Merchant accounts acceptance tests")
+public class MerchantAccountsAcceptanceTests extends AcceptanceTests {
+
+    @SneakyThrows
+    @Test
+    @DisplayName("It should get the list of merchant accounts for the given client")
+    public void itShouldGetTheListOfMerchantAccounts() {
+        ApiResponse<ListMerchantAccountsResponse> merchantAccountsResponse =
+                tlClient.merchantAccounts().listMerchantAccounts().get();
+
+        assertNotError(merchantAccountsResponse);
+    }
+
+    @SneakyThrows
+    @Test
+    @DisplayName("It should get the a merchant accounts by id")
+    public void itShouldGetAMerchantAccountById() {
+        ApiResponse<ListMerchantAccountsResponse> merchantAccountsResponse =
+                tlClient.merchantAccounts().listMerchantAccounts().get();
+        String merchantAccountId =
+                merchantAccountsResponse.getData().getItems().get(0).getId();
+        ApiResponse<truelayer.java.merchantaccounts.entities.MerchantAccount> getMerchantAccountByIdResponse =
+                tlClient.merchantAccounts()
+                        .getMerchantAccountById(merchantAccountId)
+                        .get();
+
+        assertNotError(getMerchantAccountByIdResponse);
+    }
+
+    @SneakyThrows
+    @Test
+    @DisplayName("It should get the list of transactions for a given merchant account")
+    public void itShouldGetTheListOfTransactions() {
+        ApiResponse<ListMerchantAccountsResponse> merchantAccountsResponse =
+                tlClient.merchantAccounts().listMerchantAccounts().get();
+        String merchantAccountId =
+                merchantAccountsResponse.getData().getItems().get(0).getId();
+        ApiResponse<List<Transaction>> getTransactionsResponse = tlClient.merchantAccounts()
+                .getTransactions(merchantAccountId, "2021-03-01T00:00:00.000Z", "2022-03-01T00:00:00.000Z", null)
+                .get();
+
+        assertNotError(getTransactionsResponse);
+    }
+}
